backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
dj-database-url==2.3.0
Django==5.2.1
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-filter==25.1
django-jazzmin==3.0.1
django-oauth-toolkit==3.0.1
django-seed==0.3.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Faker==37.1.0
gunicorn==20.1.0
idna==3.10
jwcrypto==1.5.6
oauthlib==3.2.2
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
rest-framework-simplejwt==0.0.2
sqlparse==0.5.3
toposort==1.10
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
whitenoise==6.4.0
